requirements.txt
requests>=2.31.0
pytest>=9.0.0
snowflake-connector-python>=4.4.0
dbt-core>=1.11.0
dbt-snowflake>=1.11.0
python-dotenv>=1.0.0
